remove app kwargs from delete_user and mass_user_update
 - manage.cmd.users: don't send updates when deleting users

the rationale for this is that while we have an app.dispatcher, it is
fully empty since we created a mock just for manage.py operations. to
have proper side-effects occouring on the actual main app, we'd need to
implement some kind of IPC, and i'm not up for that task, yet.